backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.2.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis
redis==5.2.1

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
authlib==1.3.2
httpx==0.28.0

# HTTP client (crawler)
aiohttp==3.11.10

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF reports
weasyprint==63.0

# Rate limiting
slowapi==0.1.9

# File uploads (streaming)
python-multipart==0.0.12

# CLI
typer==0.15.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0
